requirements.txt
transformers==3.0.2
numpy==1.19.1
joblib==0.16.0
flask
sentencepiece==0.1.91
urllib3==1.25.10
wikipedia
pdfminer.six